+++ clients/python/marquez_client/__init__.py
@@ -4,7 +4,7 @@
 # -*- coding: utf-8 -*-
 
 __author__ = """Marquez Project"""
-__version__ = "0.43.1"
+__version__ = "0.45.0"
 
 from marquez_client.client import MarquezClient              # noqa: F401
 from marquez_client.clients import Clients                   # noqa: F401